refactor(consolidate): extract shared domain types to consolidate/types (D3 step 1)
Move MemoryEntry, the four Consolidate*Op interfaces, ConsolidateOperation,
and RawChunkPlan into consolidate/types.ts — the cluster's dependency sink.
Re-exported from consolidate.ts so all importers (improve.ts, recombine.ts,
improve-types.ts, tests) keep resolving. Pure type move; behavior-identical.

This is the prerequisite increment for D3: leaf modules can now import these
types without a back-edge to the orchestrator. Design: docs/technical/d3-design.md.
